Adds a bootstrap-aware process director for lantern
LanternBootstrapProcessDirector monitors Lanterns output (stdout, stdin)
and fires a callback when it believes Lantern has bootstrapped. If a
timeout is configured and Lantern has not bootstrapped before it fires,
then the process director will errback instead.

This could be a useful pattern for running a variety of tests against
a single long-running process, rather than noisily bootstrapping several
times to do each test.